Tricky

+++ iotdb-core/datanode/src/main/java/org/apache/iotdb/db/schemaengine/schemaregion/attribute/update/UpdateContainer.java
@@ -39,7 +39,7 @@ public interface UpdateContainer {
   // Only this method is not synchronize called and is called by GRASS thread
   // A piece of "updateContent" won't exceed "limitBytes" in order to handle
   // thrift threshold and low bandwidth
-  default byte[] getUpdateContent(final long limitBytes) {
+  default byte[] getUpdateContent(final int limitBytes) {
     final ByteArrayOutputStream outputStream = new ByteArrayOutputStream();
     try {
       serialize(outputStream);
